#22 set alert when user want to sign out

+++ public/js/log_out.js
@@ -2,8 +2,10 @@
 
 // LOG OUT : CLEAR SESSIONSTORAGE
 function logoutAccount() {
-    sessionStorage.clear();
-    location.href = "register_view.html";
+    if (confirm("Are you sure to sign out?")) {
+        sessionStorage.clear();
+        location.href = "register_view.html";
+    }
 }
